backend/requirements.txt
langgraph==1.0.8
langgraph-cli[inmem]==0.4.12
langchain==1.2.9
langchain-openai==1.1.7
langchain-anthropic==1.3.2
langsmith==0.6.9
pydantic==2.12.5
fastapi==0.128.3
uvicorn==0.40.0
python-dotenv==1.2.1
pandas==3.0.0
numpy==2.4.2
python-multipart==0.0.22
openai==2.17.0
httpx==0.28.1
